# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface RegisterStudentRequest {
  name: string;
  phone: string;
  password: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
}

export interface UpdateStudentRequest {
  name?: string;
  phone?: string;
  password?: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
}

export interface CreateSupervisorRequest {
  name: string;
  phone: string;
  password: string;
}

export interface Student {
  id: number;
  name: string;
  phone: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
  supervisorId?: number | null;
  createdAt: string;
}

export type StudentProgressTestAttempts = { [key: string]: unknown };

export interface StudentProgress {
  id: number;
  studentId: number;
  currentWajh: number;
  correctCount: number;
  currentSession: number;
  totalStars: number;
  waitingTeacher: boolean;
  testAttempts?: StudentProgressTestAttempts;
  hashdCompleted: number;
  updatedAt?: string;
}

export type UpdateProgressRequestTestAttempts = { [key: string]: unknown };

export interface UpdateProgressRequest {
  currentWajh?: number;
  correctCount?: number;
  currentSession?: number;
  totalStars?: number;
  waitingTeacher?: boolean;
  testAttempts?: UpdateProgressRequestTestAttempts;
  hashdCompleted?: number;
}

export interface StudentWithProgress {
  id: number;
  name: string;
  phone: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
  supervisorId?: number | null;
  createdAt: string;
  progress?: StudentProgress;
}

export interface LeaderboardEntry {
  id: number;
  name: string;
  totalStars: number;
  currentWajh: number;
  rank: number;
}

export interface Supervisor {
  id: number;
  name: string;
  phone: string;
  studentCount: number;
  createdAt: string;
}

export interface SupervisorLoginResponse {
  supervisor: Supervisor;
  students: StudentWithProgress[];
}

export interface AdminStats {
  totalStudents: number;
  activeStudents: number;
  totalSupervisors: number;
  totalStars: number;
}

export interface AdminLoginResponse {
  role: string;
  stats: AdminStats;
}

export interface ResetPasswordResponse {
  newPassword: string;
}