# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalStudents: number;
  activeStudents: number;
  totalSupervisors: number;
  totalStars: number;
}