# lib/api-zod/src/generated/types/createSupervisorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSupervisorRequest {
  name: string;
  phone: string;
  password: string;
}