# lib/api-zod/src/generated/types/registerStudentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterStudentRequest {
  name: string;
  phone: string;
  password: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
}