# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  phone: string;
  country?: string;
  gender?: string;
  tajweedLevel?: string;
  supervisorId?: number | null;
  createdAt: string;
}