# lib/api-zod/src/generated/types/studentProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProgressTestAttempts } from "./studentProgressTestAttempts";

export interface StudentProgress {
  id: number;
  studentId: number;
  currentWajh: number;
  correctCount: number;
  currentSession: number;
  totalStars: number;
  waitingTeacher: boolean;
  testAttempts?: StudentProgressTestAttempts;
  hashdCompleted: number;
  updatedAt?: string;
}