# lib/api-zod/src/generated/types/supervisor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */

export interface Supervisor {
  id: number;
  name: string;
  phone: string;
  studentCount: number;
  createdAt: string;
}