# lib/api-zod/src/generated/types/supervisorLoginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentWithProgress } from "./studentWithProgress";
import type { Supervisor } from "./supervisor";

export interface SupervisorLoginResponse {
  supervisor: Supervisor;
  students: StudentWithProgress[];
}