# lib/api-zod/src/generated/types/updateProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نور البقرة API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProgressRequestTestAttempts } from "./updateProgressRequestTestAttempts";

export interface UpdateProgressRequest {
  currentWajh?: number;
  correctCount?: number;
  currentSession?: number;
  totalStars?: number;
  waitingTeacher?: boolean;
  testAttempts?: UpdateProgressRequestTestAttempts;
  hashdCompleted?: number;
}